Minimize debug  logging. Eliminate wdyr. Move code to utility modules.
