+++ client/src/components/RemoteUtilities.tsx
@@ -1,10 +1,15 @@
-import { DBCreds, RemoteDBState } from "./RemoteDBState";
+import { DBCreds, DBCredsInit, RemoteDBState } from "./RemoteDBState";
 import { CapacitorHttp, HttpOptions, HttpResponse } from '@capacitor/core';
+import { Preferences } from '@capacitor/preferences';
 import jwt_decode from 'jwt-decode';
 import { ListRow } from "./DataTypes";
+import { UUIDDoc, maxAppSupportedSchemaVersion } from "./DBSchema";
+import { DBUUIDAction, DBUUIDCheck } from "./RemoteDBState";
 import { History } from "history";
 import { urlPatternValidation, usernamePatternValidation, emailPatternValidation,
-        fullnamePatternValidation, apiConnectTimeout } from "./Utilities";
+        fullnamePatternValidation, apiConnectTimeout, isJsonString, DEFAULT_API_URL } from "./Utilities";
+import { cloneDeep, pick, keys, isEqual } from 'lodash';
+
 
 export async function navigateToFirstListID(phistory: History,remoteDBCreds: DBCreds, listRows: ListRow[]) {
     let firstListID = null;
@@ -124,3 +129,127 @@ export function errorCheckCreds({credsObj,background, creatingNewUser = false, p
         setError("Passwords do not match"); return credsCheck;}
     return credsCheck;
 }
+
+export async function checkJWT(accessJWT: string, remoteDBCreds: DBCreds) {
+    let checkResponse = {
+        JWTValid: false,
+        DBServerAvailable: true,
+        JWTExpireDate: 0
+    }
+    let response: HttpResponse | undefined;
+    checkResponse.DBServerAvailable = true;
+    const options: HttpOptions = {
+        url: String(remoteDBCreds.couchBaseURL+"/_session"),
+        method: "GET",
+        headers: { 'Content-Type': 'application/json',
+                   'Accept': 'application/json',
+                   'Authorization': 'Bearer '+ accessJWT },
+        connectTimeout: apiConnectTimeout          
+          };
+    try { response = await CapacitorHttp.get(options); }
+    catch(err) {console.log("Got error:",err); checkResponse.DBServerAvailable=false}
+    if (checkResponse.DBServerAvailable) {
+        if ((response?.status === 200) && (response.data?.userCtx?.name !== null)) {
+            let tokenInfo = getTokenInfo(accessJWT);
+            if (tokenInfo.valid) {
+                checkResponse.JWTValid = true;
+                checkResponse.JWTExpireDate = tokenInfo.expireDate;
+            }
+        } 
+    } 
+    return checkResponse;
+} 
+
+export async function checkDBUUID(db: PouchDB.Database, remoteDB: PouchDB.Database) {
+    let UUIDCheck: DBUUIDCheck = {
+        checkOK: true,
+        schemaVersion: 0,
+        dbUUIDAction: DBUUIDAction.none
+    }
+    let UUIDResults = await remoteDB.find({
+        selector: { "type": { "$eq": "dbuuid"} } })
+    let UUIDResult : null|string = null;
+    if (UUIDResults.docs.length > 0) {
+      UUIDResult = (UUIDResults.docs[0] as UUIDDoc).uuid;
+    }
+    if (UUIDResult == null) {
+      UUIDCheck.checkOK = false; UUIDCheck.dbUUIDAction = DBUUIDAction.exit_no_uuid_on_server;
+      return UUIDCheck;
+    }
+    UUIDCheck.schemaVersion = (UUIDResults.docs[0] as UUIDDoc).schemaVersion;
+    let remoteSchemaVersion = Number(UUIDCheck.schemaVersion);
+    let localDBInfo = null;
+    let localHasRecords = false;
+    let localDBUUID = null;
+    let localSchemaVersion = 0;
+    try { localDBInfo = await db.info();} catch(e) {localHasRecords=false};
+    if (localDBInfo != null && localDBInfo.doc_count > 0) { localHasRecords = true}
+    if (localHasRecords) {
+      let localDBAllDocs = null;
+      try { localDBAllDocs = await db.allDocs({include_docs: true});} catch(e) {console.log(e)};
+      localHasRecords = false;
+      if (localDBAllDocs != null) {
+        localDBAllDocs.rows.forEach(row => {
+          if ((row.doc as any).language !== "query") {
+                localHasRecords=true;
+            }
+        });
+      }
+    }
+    if (localHasRecords) {
+        let localDBFindDocs = null;
+        try { localDBFindDocs = await db.find({selector: { "type": { "$eq": "dbuuid"} }}) }
+        catch(e) {console.log(e)};
+        if ((localDBFindDocs !== null) && localDBFindDocs.docs.length === 1) {
+            localDBUUID = (localDBFindDocs.docs[0] as UUIDDoc).uuid;
+            localSchemaVersion = Number((localDBFindDocs.docs[0] as UUIDDoc).schemaVersion);
+        }
+    }
+//        console.log("maxAppSupportedVersion",maxAppSupportedSchemaVersion)
+    if (Number(UUIDCheck.schemaVersion) > maxAppSupportedSchemaVersion) {
+        UUIDCheck.checkOK = false;
+        UUIDCheck.dbUUIDAction = DBUUIDAction.exit_app_schema_mismatch;
+        return UUIDCheck;
+    }
+
+    // compare to current DBCreds one.
+    if (localDBUUID === UUIDResult) {
+        // console.log("Schema: remote:",remoteSchemaVersion," local:",localSchemaVersion);
+        if (remoteSchemaVersion > localSchemaVersion) {
+            console.log("ERROR: Remote Schema greater than local");
+            UUIDCheck.checkOK = false;
+            UUIDCheck.dbUUIDAction = DBUUIDAction.exit_local_remote_schema_mismatch;
+        }   
+        return UUIDCheck;
+    } 
+      // if current DBCreds doesn't have one, set it to the remote one.
+    if ((localDBUUID === null || localDBUUID === "" ) && !localHasRecords) {
+      return UUIDCheck;
+    }
+    UUIDCheck.checkOK = false; UUIDCheck.dbUUIDAction = DBUUIDAction.destroy_needed;
+    return UUIDCheck;
+  }
+
+  export async function  getPrefsDBCreds(curCreds: DBCreds)  {
+    let { value: credsStr } = await Preferences.get({ key: 'dbcreds'});
+    let credsObj: DBCreds = cloneDeep(DBCredsInit);
+    const credsOrigKeys = keys(credsObj);
+    if (isJsonString(String(credsStr))) {
+      credsObj=JSON.parse(String(credsStr));
+      let credsObjFiltered=pick(credsObj,['apiServerURL','couchBaseURL','database','dbUsername','email','fullName','JWT','refreshJWT','lastConflictsViewed'])
+      credsObj = credsObjFiltered;
+    }
+    const credKeys = keys(credsObj);
+    if (credsObj === null || credsObj.apiServerURL === undefined || (!isEqual(credsOrigKeys.sort(),credKeys.sort()))) {
+        credsObj = { apiServerURL: DEFAULT_API_URL,
+            couchBaseURL: "",
+            database: "",
+            dbUsername:"",
+            refreshJWT: "",
+            email: "",
+            fullName: "",
+            lastConflictsViewed: (new Date()).toISOString()
+            };
+    }
+    return credsObj;
+  }
