+++ client/src/index.tsx
@@ -1,5 +1,3 @@
-// import './wdyr'; // <--- first import
-
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import App from './App';
