+++ server/src/jwt.ts
@@ -67,12 +67,10 @@ export async function generateJWT ({ username, deviceUUID, timeString, includeRo
 
 export async function JWTMatchesUserDB(refreshJWT: string, deviceUUID: string, username: string) {
     let userDoc: any = await getUserDoc(username);
-    console.log("DEBUG: response doc: ",JSON.stringify(userDoc));
     if (userDoc.fullDoc === null) { userDoc.error=true; return false };
     if (userDoc.fullDoc.hasOwnProperty('refreshJWTs')) {
         console.log("STATUS: Refresh JWT matches the database:",userDoc.fullDoc.refreshJWTs[deviceUUID] == refreshJWT);    
     }
-    console.log("DEBUG: userDoc returned is: ",JSON.stringify(userDoc));
     if (userDoc.error) { return false;}
     if (userDoc.fullDoc.name !== username) { return false;}
     if (!userDoc.fullDoc.hasOwnProperty("refreshJWTs")) { return false;}
